Mudando bitonic sort
