--- Trabalho-Final/testeOmpBitonic.c
@@ -1,202 +1,0 @@
-#include<stdio.h>
-#include<stdlib.h>
-#include <time.h>
-
-#define MAX(A, B) (((A) > (B)) ? (A) : (B))
-#define MIN(A, B) (((A) > (B)) ? (B) : (A))
-#define UP 0
-#define DOWN 1
-#define SIZE 1024
-
-
-void bitonic_sort_seq(int start, int length, int *seq, int flag);
-void bitonic_sort_par(int start, int length, int *seq, int flag);
-void swap(int *a, int *b);
-
-int m;
-
-void GeraAleatorios(int numero[], int quantNumeros, int Limite) {
-    srand(time(NULL))   ;
-    int valor;
-
-    for (int i = 0; i < quantNumeros; i++){
-        valor = rand() % Limite;
-        numero[i] = valor;
-    }
-}
-
-main()
-{
-    int i, j;
-    int n;
-    int flag;
-    int *seq;
-
-    double startTime, elapsedTime; /* for checking/testing timing */
-    double clockZero = 0.0;
-
-    int numThreads,id;
-
-    n = SIZE;
-
-    seq = (int *) malloc (n * sizeof(int));
-
-    GeraAleatorios(seq,n,n);
-
-
-    // start
-    clock_t start = clock();
-
-    numThreads =  omp_get_max_threads();
-
-    // making sure input is okay
-    if ( n < numThreads * 2 )
-    {
-        printf("The size of the sequence is less than 2 * the number of processes.\n");
-        exit(0);
-    }
-
-    // the size of sub part
-    m = n / numThreads;
-
-    // make the sequence bitonic - part 1
-    for (i = 2; i <= m; i = 2 * i)
-    {
-#pragma omp parallel for shared(i, seq) private(j, flag)
-        for (j = 0; j < n; j += i)
-        {
-            if ((j / i) % 2 == 0)
-                flag = UP;
-            else
-                flag = DOWN;
-            bitonic_sort_seq(j, i, seq, flag);
-        }
-    }
-
-    // make the sequence bitonic - part 2
-    for (i = 2; i <= numThreads; i = 2 * i)
-    {
-        for (j = 0; j < numThreads; j += i)
-        {
-            if ((j / i) % 2 == 0)
-                flag = UP;
-            else
-                flag = DOWN;
-            bitonic_sort_par(j*m, i*m, seq, flag);
-        }
-#pragma omp parallel for shared(j)
-        for (j = 0; j < numThreads; j++)
-        {
-            if (j < i)
-                flag = UP;
-            else
-                flag = DOWN;
-            bitonic_sort_seq(j*m, m, seq, flag);
-        }
-    }
-
-    // bitonic sort
-    //bitonic_sort_par(0, n, seq, UP);
-    //bitonic_sort_seq(0, n, seq, UP);
-
-    //end
-    clock_t end = clock();
-
-    /*
-    // print a sequence
-    for (i = 0; i < n; i++){
-      printf("%d ", seq[i]);
-    }
-    printf("\n");
-    */
-
-    float seconds = (float)(end - start)/CLOCKS_PER_SEC;
-    printf("%f\n", seconds);
-    free(seq);
-}
-
-
-void bitonic_sort_seq(int start, int length, int *seq, int flag)
-{
-    int i;
-    int split_length;
-
-    if (length == 1)
-        return;
-
-    if (length % 2 !=0 )
-    {
-        printf("error\n");
-        exit(0);
-    }
-
-    split_length = length / 2;
-
-    // bitonic split
-    for (i = start; i < start + split_length; i++)
-    {
-        if (flag == UP)
-        {
-            if (seq[i] > seq[i + split_length])
-                swap(&seq[i], &seq[i + split_length]);
-        }
-        else
-        {
-            if (seq[i] < seq[i + split_length])
-                swap(&seq[i], &seq[i + split_length]);
-        }
-    }
-
-    bitonic_sort_seq(start, split_length, seq, flag);
-    bitonic_sort_seq(start + split_length, split_length, seq, flag);
-}
-
-void bitonic_sort_par(int start, int length, int *seq, int flag)
-{
-    int i;
-    int split_length;
-
-    if (length == 1)
-        return;
-
-    if (length % 2 !=0 )
-    {
-        printf("The length of a (sub)sequence is not divided by 2.\n");
-        exit(0);
-    }
-
-    split_length = length / 2;
-
-    // bitonic split
-#pragma omp parallel for shared(seq, flag, start, split_length) private(i)
-    for (i = start; i < start + split_length; i++)
-    {
-        if (flag == UP)
-        {
-            if (seq[i] > seq[i + split_length])
-                swap(&seq[i], &seq[i + split_length]);
-        }
-        else
-        {
-            if (seq[i] < seq[i + split_length])
-                swap(&seq[i], &seq[i + split_length]);
-        }
-    }
-
-    if (split_length > m)
-    {
-        // m is the size of sub part-> n/numThreads
-        bitonic_sort_par(start, split_length, seq, flag);
-        bitonic_sort_par(start + split_length, split_length, seq, flag);
-    }
-
-    return;
-}
-
-void swap(int *a, int *b)
-{
-    int t;
-    t = *a;
-    *a = *b;
-    *b = t;
-}
